update 28-06-2021 (b)
- ajout de l'approvisionnement avec confirmation de livraison
- correction des bugs

+++ src/main/java/cm/louisstark/gestock/sessions/LivraisonFacadeLocal.java
@@ -33,6 +33,6 @@ public interface LivraisonFacadeLocal {
 
     int count();
     
-    long nextId();
+    Integer nextId();
     
 }
